requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
requests
